backend/requirements.txt
# Django Backend for SocialBoosterMedia

## Requirements
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
